+++ libraries/cast/src/main/java/androidx/media3/cast/RemoteCastPlayer.java
@@ -74,6 +74,7 @@ import com.google.android.gms.cast.framework.CastSession;
 import com.google.android.gms.cast.framework.ModuleUnavailableException;
 import com.google.android.gms.cast.framework.SessionManager;
 import com.google.android.gms.cast.framework.SessionManagerListener;
+import com.google.android.gms.cast.framework.media.MediaQueue;
 import com.google.android.gms.cast.framework.media.RemoteMediaClient;
 import com.google.android.gms.cast.framework.media.RemoteMediaClient.MediaChannelResult;
 import com.google.android.gms.common.api.PendingResult;
@@ -289,6 +290,7 @@ public final class RemoteCastPlayer extends BasePlayer {
   private final Cast.Listener castListener;
 
   private final StatusListener statusListener;
+  private final MediaQueue.Callback mediaQueueCallback;
   private final SeekResultCallback seekResultCallback;
 
   // Listeners and notification.
@@ -365,6 +367,7 @@ public final class RemoteCastPlayer extends BasePlayer {
     period = new Timeline.Period();
     castListener = new CastListener();
     statusListener = new StatusListener();
+    mediaQueueCallback = new MediaQueueCallback();
     seekResultCallback = new SeekResultCallback();
     listeners =
         new ListenerSet<>(
@@ -1622,6 +1625,7 @@ public final class RemoteCastPlayer extends BasePlayer {
     if (this.remoteMediaClient != null) {
       this.remoteMediaClient.unregisterCallback(statusListener);
       this.remoteMediaClient.removeProgressListener(statusListener);
+      this.remoteMediaClient.getMediaQueue().unregisterCallback(mediaQueueCallback);
     }
     this.remoteMediaClient = remoteMediaClient;
     if (remoteMediaClient != null) {
@@ -1631,6 +1635,7 @@ public final class RemoteCastPlayer extends BasePlayer {
       if (internalSessionAvailabilityListener != null) {
         internalSessionAvailabilityListener.onCastSessionAvailable();
       }
+      remoteMediaClient.getMediaQueue().registerCallback(mediaQueueCallback);
       remoteMediaClient.registerCallback(statusListener);
       remoteMediaClient.addProgressListener(statusListener, PROGRESS_REPORT_PERIOD_MS);
       updateInternalStateAndNotifyIfChanged();
@@ -2024,6 +2029,15 @@ public final class RemoteCastPlayer extends BasePlayer {
     @Override
     public void onVolumeChanged() {
       updateDeviceVolumeAndNotifyIfChanged();
+      listeners.flushEvents();
+    }
+  }
+
+  private class MediaQueueCallback extends MediaQueue.Callback {
+
+    @Override
+    public void mediaQueueChanged() {
+      updateTimelineAndNotifyIfChanged();
       listeners.flushEvents();
     }
   }